Add an intermediate class between entity and object and group
