+++ geoh5py/shared/entity.py
@@ -20,9 +20,7 @@
 from __future__ import annotations
 
 import uuid
-import warnings
 from abc import ABC, abstractmethod
-from pathlib import Path
 from typing import TYPE_CHECKING
 
 import numpy as np
@@ -30,10 +28,7 @@ import numpy as np
 from geoh5py.shared.utils import str2uuid
 
 if TYPE_CHECKING:
-    from numpy import ndarray
-
     from .. import shared
-    from ..groups import PropertyGroup
     from ..workspace import Workspace
 
 DEFAULT_CRS = {"Code": "Unknown", "Name": "Unknown"}
@@ -61,17 +56,17 @@ class Entity(ABC):
         self._uid = (
             str2uuid(uid) if isinstance(str2uuid(uid), uuid.UUID) else uuid.uuid4()
         )
-        self._name = name
-        self._parent: Entity | None = None
-        self._children: list = []
+
         self._allow_delete = True
         self._allow_move = True
         self._allow_rename = True
-        self._partially_hidden = False
         self._clipping_ids: list[uuid.UUID] | None = None
-        self._public = True
-        self._on_file = False
         self._metadata: dict | None = None
+        self._name = name
+        self._on_file = False
+        self._parent: Entity | None = None
+        self._partially_hidden = False
+        self._public = True
 
         for attr, item in kwargs.items():
             try:
@@ -80,50 +75,6 @@ class Entity(ABC):
                 setattr(self, attr, item)
             except AttributeError:
                 continue
-
-    def add_children(self, children: list[Entity]):
-        """
-        :param children: Add a list of entities as
-            :obj:`~geoh5py.shared.entity.Entity.children`
-        """
-        if not isinstance(children, list):
-            children = [children]
-
-        for child in children:
-            if child not in self._children:
-                if not isinstance(child, Entity):
-                    raise TypeError(
-                        f"Child must be an instance of Entity, not {type(child)}"
-                    )
-                self._children.append(child)
-
-    def add_file(self, file: str):
-        """
-        Add a file to the object or group stored as bytes on a FilenameData
-
-        :param file: File name with path to import.
-        """
-        if not Path(file).is_file():
-            raise ValueError(f"Input file '{file}' does not exist.")
-
-        with open(file, "rb") as raw_binary:
-            blob = raw_binary.read()
-
-        name = Path(file).name
-        attributes = {
-            "name": name,
-            "file_name": name,
-            "association": "OBJECT",
-            "parent": self,
-            "values": blob,
-        }
-        entity_type = {"name": "UserFiles", "primitive_type": "FILENAME"}
-
-        file_data = self.workspace.create_entity(
-            None, entity=attributes, entity_type=entity_type
-        )
-
-        return file_data
 
     @property
     def allow_delete(self) -> bool:
@@ -170,58 +121,11 @@ class Entity(ABC):
         return self._attribute_map
 
     @property
-    def children(self):
-        """
-        :obj:`list` Children entities in the workspace tree
-        """
-        return self._children
-
-    @property
     def clipping_ids(self) -> list[uuid.UUID] | None:
         """
         List of clipping uuids
         """
         return self._clipping_ids
-
-    @abstractmethod
-    def mask_by_extent(
-        self, extent: np.ndarray, inverse: bool = False
-    ) -> np.ndarray | None:
-        """
-        Get a mask array from coordinate extent.
-
-        :param extent: Bounding box extent coordinates defined by either:
-            - obj:`numpy.ndarray` of shape (2, 3)
-            3D coordinate: [[west, south, bottom], [east, north, top]]
-            - obj:`numpy.ndarray` of shape (2, 2)
-            Horizontal coordinates: [[west, south], [east, north]].
-        :param inverse: Return the complement of the mask extent. Default to False
-
-        :return: Array of bool defining the vertices or cell centers
-            within the mask extent, or None if no intersection.
-        """
-
-    @classmethod
-    def create(cls, workspace, **kwargs):
-        """
-        Function to create an entity.
-
-        :param workspace: Workspace to be added to.
-        :param kwargs: List of keyword arguments defining the properties of a class.
-
-        :return entity: Registered Entity to the workspace.
-        """
-        entity_type_kwargs = (
-            {"entity_type": {"uid": kwargs["entity_type_uid"]}}
-            if "entity_type_uid" in kwargs
-            else {}
-        )
-        entity_kwargs = {"entity": kwargs}
-        new_object = workspace.create_entity(
-            cls,
-            **{**entity_kwargs, **entity_type_kwargs},
-        )
-        return new_object
 
     @property
     def coordinate_reference_system(self) -> dict:
@@ -263,63 +167,27 @@ class Entity(ABC):
 
         self.metadata = metadata
 
-    @abstractmethod
-    def copy(
-        self,
-        parent=None,
-        copy_children: bool = True,
-        clear_cache: bool = False,
-        mask: np.ndarray | None = None,
-        **kwargs,
-    ):
+    @classmethod
+    def create(cls, workspace, **kwargs):
         """
-        Function to copy an entity to a different parent entity.
+        Function to create an entity.
 
-        :param parent: Target parent to copy the entity under. Copied to current
-            :obj:`~geoh5py.shared.entity.Entity.parent` if None.
-        :param copy_children: (Optional) Create copies of all children entities along with it.
-        :param clear_cache: Clear array attributes after copy to minimize the
-            memory footprint of the workspace.
-        :param mask: Array of indices to sub-sample the input entity.
-        :param kwargs: Additional keyword arguments to pass to the copy constructor.
+        :param workspace: Workspace to be added to.
+        :param kwargs: List of keyword arguments defining the properties of a class.
 
         :return entity: Registered Entity to the workspace.
         """
-
-    def copy_from_extent(
-        self,
-        extent: ndarray,
-        parent=None,
-        copy_children: bool = True,
-        clear_cache: bool = False,
-        inverse: bool = False,
-        **kwargs,
-    ) -> Entity | None:
-        """
-        Function to copy an entity to a different parent entity.
-
-        :param extent: Bounding box extent requested for the input entity, as supplied for
-            :func:`~geoh5py.shared.entity.Entity.mask_by_extent`.
-        :param parent: Target parent to copy the entity under. Copied to current
-            :obj:`~geoh5py.shared.entity.Entity.parent` if None.
-        :param copy_children: (Optional) Create copies of all children entities along with it.
-        :param clear_cache: Clear array attributes after copy.
-        :param inverse: Keep the inverse (clip) of the extent selection.
-        :param kwargs: Additional keyword arguments to pass to the copy constructor.
-
-        :return entity: Registered Entity to the workspace.
-        """
-        indices = self.mask_by_extent(extent, inverse=inverse)
-        if indices is None:
-            return None
-
-        return self.copy(
-            parent=parent,
-            copy_children=copy_children,
-            clear_cache=clear_cache,
-            mask=indices,
-            **kwargs,
+        entity_type_kwargs = (
+            {"entity_type": {"uid": kwargs["entity_type_uid"]}}
+            if "entity_type_uid" in kwargs
+            else {}
         )
+        entity_kwargs = {"entity": kwargs}
+        new_object = workspace.create_entity(
+            cls,
+            **{**entity_kwargs, **entity_type_kwargs},
+        )
+        return new_object
 
     @property
     @abstractmethod
@@ -336,36 +204,23 @@ class Entity(ABC):
         #  (possibly it has to be abstract with different implementations per Entity type)
         return name
 
-    def get_entity(self, name: str | uuid.UUID) -> list[Entity | None]:
+    @abstractmethod
+    def mask_by_extent(
+        self, extent: np.ndarray, inverse: bool = False
+    ) -> np.ndarray | None:
         """
-        Get a child :obj:`~geoh5py.data.data.Data` by name.
+        Get a mask array from coordinate extent.
 
-        :param name: Name of the target child data
-        :param entity_type: Sub-select entities based on type.
-        :return: A list of children Data objects
+        :param extent: Bounding box extent coordinates defined by either:
+            - obj:`numpy.ndarray` of shape (2, 3)
+            3D coordinate: [[west, south, bottom], [east, north, top]]
+            - obj:`numpy.ndarray` of shape (2, 2)
+            Horizontal coordinates: [[west, south], [east, north]].
+        :param inverse: Return the complement of the mask extent. Default to False
+
+        :return: Array of bool defining the vertices or cell centers
+            within the mask extent, or None if no intersection.
         """
-
-        if isinstance(name, uuid.UUID):
-            entity_list = [child for child in self.children if child.uid == name]
-        else:
-            entity_list = [child for child in self.children if child.name == name]
-
-        if not entity_list:
-            return [None]
-
-        return entity_list
-
-    def get_entity_list(self, entity_type=ABC) -> list[str]:
-        """
-        Get a list of names of all children :obj:`~geoh5py.data.data.Data`.
-
-        :param entity_type: Option to sub-select based on type.
-        :return: List of names of data associated with the object.
-        """
-        name_list = [
-            child.name for child in self.children if isinstance(child, entity_type)
-        ]
-        return sorted(name_list)
 
     @property
     def metadata(self) -> dict | None:
@@ -418,11 +273,15 @@ class Entity(ABC):
     def parent(self, parent: shared.Entity):
         current_parent = self._parent
 
-        if hasattr(parent, "add_children"):
+        if hasattr(parent, "add_children") and hasattr(parent, "remove_children"):
             parent.add_children([self])
             self._parent = parent
 
-            if current_parent is not None and current_parent != self._parent:
+            if (
+                current_parent is not None
+                and current_parent != self._parent
+                and hasattr(current_parent, "remove_children")
+            ):
                 current_parent.remove_children([self])
                 self.workspace.save_entity(self)
 
@@ -450,60 +309,6 @@ class Entity(ABC):
     def public(self, value: bool):
         self._public = value
         self.workspace.update_attribute(self, "attributes")
-
-    def reference_to_uid(
-        self, value: Entity | PropertyGroup | str | uuid.UUID
-    ) -> list[uuid.UUID]:
-        """
-        General entity reference translation.
-
-        :param value: Either an `Entity`, string or uuid
-
-        :return: List of unique identifier associated with the input reference.
-        """
-        children_uid = [child.uid for child in self.children]
-        if hasattr(value, "uid"):
-            uid = [value.uid]
-        elif isinstance(value, str):
-            uid = [
-                obj.uid
-                for obj in self.workspace.get_entity(value)
-                if (obj is not None) and (obj.uid in children_uid)
-            ]
-        elif isinstance(value, uuid.UUID):
-            uid = [value]
-
-        return uid
-
-    def remove_children(self, children: list[shared.Entity] | list[PropertyGroup]):
-        """
-        Remove children from the list of children entities.
-
-        :param children: List of entities
-
-        .. warning::
-            Removing a child entity without re-assigning it to a different
-            parent may cause it to become inactive. Inactive entities are removed
-            from the workspace by
-            :func:`~geoh5py.shared.weakref_utils.remove_none_referents`.
-        """
-        if not isinstance(children, list):
-            children = [children]
-
-        self._children = [child for child in self._children if child not in children]
-        self.workspace.remove_children(self, children)
-
-    def save(self, add_children: bool = True):
-        """
-        Alias method of :func:`~geoh5py.workspace.Workspace.save_entity`.
-        WILL BE DEPRECATED AS ENTITIES ARE ALWAYS AUTOMATICALLY UPDATED.
-        :param add_children: Option to also save the children.
-        """
-        warnings.warn(
-            "Entity.save() is deprecated and will be removed in next versions.",
-            DeprecationWarning,
-        )
-        return self.workspace.save_entity(self, add_children=add_children)
 
     @property
     def uid(self) -> uuid.UUID:
